+++ src/io_command.cpp
@@ -1,6 +1,6 @@
 #include "io_command.h"
 
-IO_Command::IO_Command(float issueTime, char type) : Command(issueTime, type) {
+IO_Command::IO_Command(double issueTime, char type) : Command(issueTime, type) {
 
 }
