+++ src/trim_command.cpp
@@ -1,6 +1,6 @@
 #include "trim_command.h"
 
-Trim_Command::Trim_Command(float issueTime) : Command(issueTime, 'T')	{
+Trim_Command::Trim_Command(double issueTime) : Command(issueTime, 'T')	{
 
 }
